refactor to use sqlx macros

+++ src-tauri/src/ai/rag/engines/settings.rs
@@ -2,7 +2,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 // Supporting enums
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub enum RAGChunkSelectionMethod {
     #[serde(rename = "weight")]
     Weight,
@@ -10,7 +10,7 @@ pub enum RAGChunkSelectionMethod {
     Vector,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub enum RAGSimpleGraphQueryMode {
     #[serde(rename = "local")]
     Local,
@@ -27,20 +27,20 @@ pub enum RAGSimpleGraphQueryMode {
 }
 
 // Vector Engine Settings
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleVectorIndexingSettings {
-    pub chunk_token_size: Option<usize>,
-    pub chunk_overlap_token_size: Option<usize>,
+    pub chunk_token_size: Option<i32>,
+    pub chunk_overlap_token_size: Option<i32>,
     pub cosine_better_than_threshold: Option<f32>,
 }
 
 impl RAGSimpleVectorIndexingSettings {
     pub fn chunk_token_size(&self) -> usize {
-        self.chunk_token_size.unwrap_or(1200) // CHUNK_SIZE
+        self.chunk_token_size.unwrap_or(1200) as usize // CHUNK_SIZE
     }
 
     pub fn chunk_overlap_token_size(&self) -> usize {
-        self.chunk_overlap_token_size.unwrap_or(100) // CHUNK_OVERLAP_SIZE
+        self.chunk_overlap_token_size.unwrap_or(100) as usize // CHUNK_OVERLAP_SIZE
     }
 
     pub fn cosine_better_than_threshold(&self) -> f32 {
@@ -48,13 +48,13 @@ impl RAGSimpleVectorIndexingSettings {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleVectorQueryingSettings {
-    pub top_k: Option<usize>,
-    pub chunk_top_k: Option<usize>,
+    pub top_k: Option<i32>,
+    pub chunk_top_k: Option<i32>,
     pub similarity_threshold: Option<f32>,
-    pub related_chunk_number: Option<usize>,
-    pub max_total_tokens: Option<usize>,
+    pub related_chunk_number: Option<i32>,
+    pub max_total_tokens: Option<i32>,
     pub chunk_selection_method: Option<RAGChunkSelectionMethod>,
     pub user_prompt: Option<String>,
     pub enable_rerank: Option<bool>,
@@ -63,11 +63,11 @@ pub struct RAGSimpleVectorQueryingSettings {
 
 impl RAGSimpleVectorQueryingSettings {
     pub fn top_k(&self) -> usize {
-        self.top_k.unwrap_or(40) // DEFAULT_TOP_K
+        self.top_k.unwrap_or(40) as usize // DEFAULT_TOP_K
     }
 
     pub fn chunk_top_k(&self) -> usize {
-        self.chunk_top_k.unwrap_or(20) // DEFAULT_CHUNK_TOP_K
+        self.chunk_top_k.unwrap_or(20)  as usize // DEFAULT_CHUNK_TOP_K
     }
 
     pub fn similarity_threshold(&self) -> f32 {
@@ -75,11 +75,11 @@ impl RAGSimpleVectorQueryingSettings {
     }
 
     pub fn related_chunk_number(&self) -> usize {
-        self.related_chunk_number.unwrap_or(5) // DEFAULT_RELATED_CHUNK_NUMBER
+        self.related_chunk_number.unwrap_or(5)  as usize // DEFAULT_RELATED_CHUNK_NUMBER
     }
 
     pub fn max_total_tokens(&self) -> usize {
-        self.max_total_tokens.unwrap_or(30000) // DEFAULT_MAX_TOTAL_TOKENS
+        self.max_total_tokens.unwrap_or(30000)  as usize // DEFAULT_MAX_TOTAL_TOKENS
     }
 
     pub fn chunk_selection_method(&self) -> RAGChunkSelectionMethod {
@@ -97,7 +97,7 @@ impl RAGSimpleVectorQueryingSettings {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleVectorEngineSettings {
     pub indexing: Option<RAGSimpleVectorIndexingSettings>,
     pub querying: Option<RAGSimpleVectorQueryingSettings>,
@@ -132,41 +132,41 @@ impl RAGSimpleVectorEngineSettings {
 }
 
 // Graph Engine Settings
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleGraphIndexingSettings {
-    pub chunk_token_size: Option<usize>,
-    pub chunk_overlap_token_size: Option<usize>,
-    pub entity_extract_max_gleaning: Option<usize>,
-    pub force_llm_summary_on_merge: Option<usize>,
-    pub max_graph_nodes: Option<usize>,
-    pub summary_max_tokens: Option<usize>,
+    pub chunk_token_size: Option<i32>,
+    pub chunk_overlap_token_size: Option<i32>,
+    pub entity_extract_max_gleaning: Option<i32>,
+    pub force_llm_summary_on_merge: Option<i32>,
+    pub max_graph_nodes: Option<i32>,
+    pub summary_max_tokens: Option<i32>,
     pub entity_types: Option<Vec<String>>,
     pub extraction_language: Option<String>,
 }
 
 impl RAGSimpleGraphIndexingSettings {
     pub fn chunk_token_size(&self) -> usize {
-        self.chunk_token_size.unwrap_or(1200) // CHUNK_SIZE
+        self.chunk_token_size.unwrap_or(1200) as usize // CHUNK_SIZE
     }
 
     pub fn chunk_overlap_token_size(&self) -> usize {
-        self.chunk_overlap_token_size.unwrap_or(100) // CHUNK_OVERLAP_SIZE
+        self.chunk_overlap_token_size.unwrap_or(100)  as usize // CHUNK_OVERLAP_SIZE
     }
 
     pub fn entity_extract_max_gleaning(&self) -> usize {
-        self.entity_extract_max_gleaning.unwrap_or(1) // DEFAULT_MAX_GLEANING
+        self.entity_extract_max_gleaning.unwrap_or(1)  as usize // DEFAULT_MAX_GLEANING
     }
 
     pub fn force_llm_summary_on_merge(&self) -> usize {
-        self.force_llm_summary_on_merge.unwrap_or(4) // DEFAULT_FORCE_LLM_SUMMARY_ON_MERGE
+        self.force_llm_summary_on_merge.unwrap_or(4)  as usize // DEFAULT_FORCE_LLM_SUMMARY_ON_MERGE
     }
 
     pub fn max_graph_nodes(&self) -> usize {
-        self.max_graph_nodes.unwrap_or(1000) // DEFAULT_MAX_GRAPH_NODES
+        self.max_graph_nodes.unwrap_or(1000)  as usize // DEFAULT_MAX_GRAPH_NODES
     }
 
     pub fn summary_max_tokens(&self) -> usize {
-        self.summary_max_tokens.unwrap_or(30000) // DEFAULT_SUMMARY_MAX_TOKENS
+        self.summary_max_tokens.unwrap_or(30000)  as usize // DEFAULT_SUMMARY_MAX_TOKENS
     }
 
     pub fn entity_types(&self) -> Vec<String> {
@@ -188,15 +188,15 @@ impl RAGSimpleGraphIndexingSettings {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleGraphQueryingSettings {
-    pub max_entity_tokens: Option<usize>,
-    pub max_relation_tokens: Option<usize>,
-    pub max_total_tokens: Option<usize>,
-    pub max_graph_nodes_per_query: Option<usize>,
-    pub top_k: Option<usize>,
-    pub chunk_top_k: Option<usize>,
-    pub related_chunk_number: Option<usize>,
+    pub max_entity_tokens: Option<i32>,
+    pub max_relation_tokens: Option<i32>,
+    pub max_total_tokens: Option<i32>,
+    pub max_graph_nodes_per_query: Option<i32>,
+    pub top_k: Option<i32>,
+    pub chunk_top_k: Option<i32>,
+    pub related_chunk_number: Option<i32>,
     pub query_mode: Option<RAGSimpleGraphQueryMode>,
     pub chunk_selection_method: Option<RAGChunkSelectionMethod>,
     pub user_prompt: Option<String>,
@@ -206,31 +206,31 @@ pub struct RAGSimpleGraphQueryingSettings {
 
 impl RAGSimpleGraphQueryingSettings {
     pub fn max_entity_tokens(&self) -> usize {
-        self.max_entity_tokens.unwrap_or(6000) // DEFAULT_MAX_ENTITY_TOKENS
+        self.max_entity_tokens.unwrap_or(6000) as usize // DEFAULT_MAX_ENTITY_TOKENS
     }
 
     pub fn max_relation_tokens(&self) -> usize {
-        self.max_relation_tokens.unwrap_or(8000) // DEFAULT_MAX_RELATION_TOKENS
+        self.max_relation_tokens.unwrap_or(8000) as usize // DEFAULT_MAX_RELATION_TOKENS
     }
 
     pub fn max_total_tokens(&self) -> usize {
-        self.max_total_tokens.unwrap_or(30000) // DEFAULT_MAX_TOTAL_TOKENS
+        self.max_total_tokens.unwrap_or(30000) as usize // DEFAULT_MAX_TOTAL_TOKENS
     }
 
     pub fn max_graph_nodes_per_query(&self) -> usize {
-        self.max_graph_nodes_per_query.unwrap_or(1000) // DEFAULT_MAX_GRAPH_NODES
+        self.max_graph_nodes_per_query.unwrap_or(1000) as usize // DEFAULT_MAX_GRAPH_NODES
     }
 
     pub fn top_k(&self) -> usize {
-        self.top_k.unwrap_or(40) // DEFAULT_TOP_K
+        self.top_k.unwrap_or(40) as usize // DEFAULT_TOP_K
     }
 
     pub fn chunk_top_k(&self) -> usize {
-        self.chunk_top_k.unwrap_or(20) // DEFAULT_CHUNK_TOP_K
+        self.chunk_top_k.unwrap_or(20) as usize // DEFAULT_CHUNK_TOP_K
     }
 
     pub fn related_chunk_number(&self) -> usize {
-        self.related_chunk_number.unwrap_or(5) // DEFAULT_RELATED_CHUNK_NUMBER
+        self.related_chunk_number.unwrap_or(5) as usize // DEFAULT_RELATED_CHUNK_NUMBER
     }
 
     pub fn query_mode(&self) -> RAGSimpleGraphQueryMode {
@@ -254,7 +254,7 @@ impl RAGSimpleGraphQueryingSettings {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, sqlx::Type)]
 pub struct RAGSimpleGraphEngineSettings {
     pub indexing: Option<RAGSimpleGraphIndexingSettings>,
     pub querying: Option<RAGSimpleGraphQueryingSettings>,
